# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Project Tasks by Team41",
        "title": "Tasks API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/task",
    "paths": {
        "/task/": {
            "get": {
                "description": "List of tasks",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "List of tasks",
                "parameters": [
                    {
                        "description": "List",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/task.ListRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            },
            "post": {
                "description": "Create task",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create task",
                "parameters": [
                    {
                        "description": "New Task",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/task.CreateRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        },
        "/task/{id}/": {
            "get": {
                "description": "Read task",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Read task",
                "parameters": [
                    {
                        "description": "Read",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/task.ReadRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            },
            "put": {
                "description": "Update task",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update task",
                "parameters": [
                    {
                        "description": "Task",
                        "name": "task",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Task"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            },
            "delete": {
                "description": "Delete task",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Delete task",
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "400": {
                        "description": ""
                    },
                    "500": {
                        "description": ""
                    }
                }
            }
        }
    },
    "definitions": {
        "models.Task": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2021-05-25T00:53:16.535668Z"
                },
                "creator_email": {
                    "type": "string",
                    "example": "creator@task.com"
                },
                "description": {
                    "type": "string",
                    "example": "Description task"
                },
                "ended_at": {
                    "type": "string",
                    "format": "date-time",
                    "example": "2021-05-25T00:53:16.535668Z"
                },
                "id": {
                    "type": "integer",
                    "example": 123
                },
                "name": {
                    "type": "string",
                    "example": "Test task"
                }
            }
        },
        "task.CreateRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "Test description"
                },
                "name": {
                    "type": "string",
                    "example": "Test name"
                },
                "participant_emails": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "task.ListRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string",
                    "example": "test@test.org"
                }
            }
        },
        "task.ReadRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Test task"
                }
            }
        }
    }
}